backend/auth_service/requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
flask_bcrypt==1.0.1
flask_jwt_extended==4.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
